move TrustedPeer to a separate module
