+++ jormungandr/src/settings/start/config.rs
@@ -1,7 +1,10 @@
 use crate::{
     network::p2p::{topic, Id, PolicyConfig},
-    settings::logging::{LogFormat, LogOutput},
-    settings::LOG_FILTER_LEVEL_POSSIBLE_VALUES,
+    settings::{
+        logging::{LogFormat, LogOutput},
+        start::trusted_peer::*,
+        LOG_FILTER_LEVEL_POSSIBLE_VALUES,
+    },
 };
 use jormungandr_lib::{interfaces::Mempool, time::Duration};
 use poldercast;
@@ -110,13 +113,6 @@ pub struct P2pConfig {
     pub max_unreachable_nodes_to_connect_per_event: Option<usize>,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub struct TrustedPeer {
-    pub address: TrustedAddress,
-    pub id: Id,
-}
-
 #[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
 #[serde(deny_unknown_fields)]
 pub struct Leadership {
@@ -129,9 +125,6 @@ pub struct Leadership {
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct Address(pub poldercast::Address);
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct TrustedAddress(pub multiaddr::Multiaddr);
 
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
 pub struct Topic(pub poldercast::Topic);
@@ -185,29 +178,6 @@ impl Default for Leadership {
     }
 }
 
-impl std::str::FromStr for TrustedPeer {
-    type Err = String;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let mut split = s.split('@');
-
-        let address = if let Some(address) = split.next() {
-            multiaddr::Multiaddr::from_bytes(address.as_bytes().iter().cloned().collect())
-                .map(TrustedAddress)
-                .map_err(|e| e.to_string())?
-        } else {
-            return Err("Missing address component".to_owned());
-        };
-
-        let id = if let Some(id) = split.next() {
-            id.parse::<Id>().map_err(|e| e.to_string())?
-        } else {
-            return Err("Missing id component".to_owned());
-        };
-
-        Ok(TrustedPeer { address, id })
-    }
-}
-
 impl Address {
     pub fn to_socketaddr(&self) -> Option<SocketAddr> {
         self.0.to_socketaddr()
@@ -221,86 +191,13 @@ custom_error! {pub AddressError
     UnsupportedProtocol = "the provided protocol is unsupported, please use one of ip4/ip6/dns4/dns6",
 }
 
-impl TrustedAddress {
-    pub fn to_addresses(&self) -> Result<Vec<Address>, AddressError> {
-        use multiaddr::AddrComponent;
-        use std::{iter::FromIterator, net::ToSocketAddrs};
-
-        let mut components = self.0.iter();
-        let protocol = components.next();
-
-        if let Some(AddrComponent::IP4(_)) | Some(AddrComponent::IP6(_)) = protocol {
-            return Ok(vec![Address(
-                poldercast::Address::new(self.0.clone()).unwrap(),
-            )]);
-        }
-
-        let port = match components.next() {
-            Some(AddrComponent::TCP(port)) => port,
-            _ => return Err(AddressError::NoPortSpecified),
-        };
-
-        let addresses: Vec<AddrComponent> = match protocol {
-            Some(AddrComponent::DNS4(fqdn)) => format!("{}:{}", fqdn, port)
-                .to_socket_addrs()
-                .map_err(|e| AddressError::DnsLookupError { source: e })?
-                .into_iter()
-                .filter_map(|r| match r {
-                    SocketAddr::V4(addr) => Some(AddrComponent::IP4(*addr.ip())),
-                    _ => None,
-                })
-                .collect(),
-            Some(AddrComponent::DNS6(fqdn)) => format!("{}:{}", fqdn, port)
-                .to_socket_addrs()
-                .map_err(|e| AddressError::DnsLookupError { source: e })?
-                .into_iter()
-                .filter_map(|r| match r {
-                    SocketAddr::V6(addr) => Some(AddrComponent::IP6(*addr.ip())),
-                    _ => None,
-                })
-                .collect(),
-            _ => return Err(AddressError::UnsupportedProtocol),
-        };
-
-        if addresses.is_empty() {
-            return Err(AddressError::NoAppropriateDNSFound);
-        }
-
-        let addresses = addresses
-            .into_iter()
-            .map(|addr| {
-                let new_components = vec![addr, AddrComponent::TCP(port)];
-                let new_multiaddr = multiaddr::Multiaddr::from_iter(new_components.into_iter());
-                Address(poldercast::Address::new(new_multiaddr).unwrap())
-            })
-            .collect();
-
-        Ok(addresses)
-    }
-}
-
 impl std::fmt::Display for Address {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
         self.0.fmt(f)
     }
 }
 
-impl std::fmt::Display for TrustedAddress {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
 impl Serialize for Address {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        serializer.serialize_str(&format!("{}", self.0))
-    }
-}
-
-impl Serialize for TrustedAddress {
     fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
     where
         S: Serializer,
@@ -362,34 +259,6 @@ impl<'de> Deserialize<'de> for Address {
             }
         }
         deserializer.deserialize_str(AddressVisitor)
-    }
-}
-
-impl<'de> Deserialize<'de> for TrustedAddress {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        struct TrustedAddressVisitor;
-        impl<'de> Visitor<'de> for TrustedAddressVisitor {
-            type Value = TrustedAddress;
-
-            fn expecting(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
-                write!(fmt, "Multiaddr (example: /ip4/192.168.0.1/tcp/443)")
-            }
-
-            fn visit_str<'a, E>(self, v: &'a str) -> std::result::Result<Self::Value, E>
-            where
-                E: serde::de::Error,
-            {
-                use serde::de::Unexpected;
-                match v.parse() {
-                    Err(_err) => Err(E::invalid_value(Unexpected::Str(v), &self)),
-                    Ok(addr) => Ok(TrustedAddress(addr)),
-                }
-            }
-        }
-        deserializer.deserialize_str(TrustedAddressVisitor)
     }
 }
